Subdivide Parser.Expressions.Literals at creation/call boundary (#1361)
Pure move: Parser.Expressions.Literals.cs (2,858 lines, over the ADR-0150
satellite cap) splits at the ParseCallExpression boundary into
Parser.Expressions.Creation.cs (map/anonymous-class/array creation,
initializers, generic-call lookahead, 1,375 lines) and the remaining
call/keyword-expression/literal band (1,498 lines). Verified by
sorted-line conservation.
